+++ crates/core/executor/src/events/precompiles/fptower.rs
@@ -24,7 +24,7 @@ pub struct FpOpEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The pointer to the x operand.
     pub x_ptr: u32,
     /// The x operand.
@@ -51,7 +51,7 @@ pub struct Fp2AddSubEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The operation to perform.
     pub op: FieldOperation,
     /// The pointer to the x operand.
@@ -76,7 +76,7 @@ pub struct Fp2MulEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The pointer to the x operand.
     pub x_ptr: u32,
     /// The x operand.
